refactor: 승패 계산 로직 ParticipantResult에서 Participant로 이동
