+++ src/main/java/blackjack/model/game/ParticipantResult.java
@@ -15,26 +15,6 @@ public enum ParticipantResult {
         this.detail = detail;
     }
 
-    public static ParticipantResult of(Dealer dealer, Participant participant) {
-        if (participant.isBust()) {
-            return LOSE;
-        }
-        if (dealer.isBust()) {
-            return WIN;
-        }
-
-        int dealerPoint = dealer.calculatePoint();
-        int participantPoint = participant.calculatePoint();
-
-        if (dealerPoint > participantPoint) {
-            return LOSE;
-        }
-        if (dealerPoint < participantPoint) {
-            return WIN;
-        }
-        return DRAW;
-    }
-
     public String getDetail() {
         return detail;
     }
